first try

+++ ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: cburns <cburns@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/06 04:26:52 by cburns            #+#    #+#             */
-/*   Updated: 2019/09/06 07:25:18 by cburns           ###   ########.fr       */
+/*   Updated: 2019/09/14 20:57:35 by cburns           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,5 +19,5 @@ int		ft_strcmp(char const *s1, char const *s2)
 	i = 0;
 	while (s1[i] == s2[i] && (s1[i]) && (s2[i]))
 		i++;
-	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
+	return ((int)((unsigned char)s1[i] - (unsigned char)s2[i]));
 }
